backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
twilio
opencv-python
pymongo
sendgrid
python-multipart
ultralytics
numpy
pydantic
python-jose[cryptography]
bcrypt
pydantic[email]
passlib